// src/types.ts
// Array of generation methods.
export const generationMethods = [
  'string',
  'bytes',
  'base32',
  'base62',
  'uuidv4',
  'uuidv5',
] as const;

/**
 * Generation methods for key/token generation.
 */
 export type GenerationMethod = (typeof generationMethods)[number];

/**
 * API key base generation options.
 */
interface __BaseGenerationOptions {
  method?: GenerationMethod;
  prefix?: string;
  batch?: number;
}

/**
 * Generation options for the "string" method.
 */
export interface StringGenerationOptions extends __BaseGenerationOptions {
  min?: number;
  max?: number;
  length?: number;
  pool?: string;
}

/**
 * Generation options for the "bytes" method.
 */
export interface BytesGenerationOptions extends __BaseGenerationOptions {
  min?: number;
  max?: number;
  length?: number;
}

/**
 * Generation options for the "base32" method.
 */
export interface Base32GenerationOptions extends __BaseGenerationOptions {
  dashes?: boolean;
}

/**
 * Generation options for the "base62" method.
 */
export type Base62GenerationOptions = __BaseGenerationOptions;

/**
 * Generation options for the "uuidv4" method.
 */
export interface UuidV4GenerationOptions extends __BaseGenerationOptions {
  dashes?: boolean;
}

/**
 * Generation options for the "uuidv5" method.
 */
export interface UuidV5GenerationOptions extends __BaseGenerationOptions {
  name?: string;
  namespace?: string;
  dashes?: boolean;
}

/**
 * API key/token generation options.
 */
export type GenerationOptions =
  | StringGenerationOptions
  | BytesGenerationOptions
  | Base32GenerationOptions
  | Base62GenerationOptions
  | UuidV4GenerationOptions
  | UuidV5GenerationOptions;

 /**
  * Default generation options.
  */
export interface DefaultGenerationOptions {
  min?: number;
  max?: number;
  pool?: string;
  dashes?: boolean;
}

/**
 * Results from the API key/token generation.
 */
export type ApiKeyResults = string | string[];